dnetview: create new msg log HashMap and render it
messages are perodically cleared on the p2p side so dnetview must keep
track of messages.

because View has no permanence, we create a HashMap<String, Vec<(String,
String)> in Model to store the message log.

this allows us to keep track of the message log for every connection.
the values are copied into View before each render.