Add out-of-memory crash to corpora
This input caused an out-of-memory bug. I haven't figured out why yet
but it's good to keep this in the corpus for future testing.

Something interesting: the fuzzer brute-forced the MAGIC BYTES for the
binary on its own! This file is a good starting point for future fuzzing
because it contains the magic bytes: subsequent runs can start from here
instead of needing to "find" the magic bytes via brute forcing every
time