docs: revert darkirc to ircd
This revert was perfomed using:
git show de294ad28c8572fac0f8fa345bb30b2383fb85ce -- doc/src/misc/ | git apply -R --reject -

--reject was used due to renaming of files and some other issues that
prevented a clean reverse patch or resolution via the -3 flag for
git-apply.

I used :%s/darkirc/ircd/g in the conflicting file and this commit is the
result of these steps combined.

Note: the following section of text is "orphaned" because I'm not sure
whether it applies to ircd or only darkirc. We should return to it and
add it back to the docs when darkirc is functioning properly.

+Whenever you edit `darkirc_config.toml` file and if you have your
+`darkirc` daemon running you don't need to restart it to reload the
+config, you just need to send a SIGHUP signal to it for the changes to
+reflect, one of the easiest ways to do that is:
+
+```shell
+% kill -SIGHUP $(pidof darkirc)
 ```