zkas: Fix compiler panic when semicolons are missing (#220)
* zkas: Add error for missing semicolons

Fixed a parser bug where it would pass invalid statements to the
compiler which resulted in a panic.

Added error handling in the parser when a semicolon is missing. The
error message suggest to the dev to add a semicolon.

Also added more detail to the unreachable()! statement in the compiler
so that a darkfi dev can see what the erroneous StatementType was that
caused the panic.

Example errror cases:

```
<statement> <statement>;
```

and
```
<statement>
<statement>;
```

To test: remove a semicolon from example/simple.zk and run ./zkas
example/simple.zk

* add code comment to explain

---------

Co-authored-by: y <y>