[bug] native_range_check: only constrains multiples of WINDOW_SIZE
In the case where the final chunk is less than WINDOW_SIZE, the
witness generation code manually pads it with zero bits.

However, these padding bits are not constrained to be zero in the
circuit. This means that the final chunk is actually allowed to be
any WINDOW_SIZE-bit value.