[net] Make Tor debugging more verbose
Add tor-error crate in order to surface more verbose debug information
when establishing connections to Tor via arti-client. (See links
including in code comments.)

Arti low-level errors are now reported at WARN level in our debug logs.

In order to implement this feature, I had to remove our patch directives
in Cargo.toml. Since we are using the up-to-date version of Arti anyway,
this (hopefully) should not cause issues.

If a revert is necessary to fix our version of Arti to some specific
commit, it is likley Tor-error will break.