[net] Add debug messages when receiving old Events
Added debug statements when an EventPut occurs for an event with
a timestamp that is older than the Genesis event of the EventGraph.

Ideally this shouldn't happen but could theoretically occur if a node
does not properly prune its DAG.

In order to do this, `genesis` and `days_rotation` were added as fields
to EventGraph and 'getter' methods were added to retrieve these values.