sdk-util: Add size limits on data; comment code
Change `data.len() as u32` to `u32::try_from`. This will return an error
in the case where the length of the data exceeds u32. (Using an `as`
conversion will cause the value to overflow and write the wrong amount
of data).

Add a new ContractError enum `DataTooLarge` to handle this case.

Add a similar check to the `parse_ret` auxiliary function to ensure the
ret i64 fits into the u32 type.

Add code comments to the utility functions to explain what they do.