channel: expand ChannelInfo to include resolve_addr and connect_addr
This paves the way for DEP 0001: https://darkrenaissance.github.io/darkfi/dep/0001.html

We retain the same method for returning the channel's address, channel.address() but modify it so that it adapts depending on whether this channel is inbound or outbound. The usage remains the same and the return value is equivalent.