outbound_session: remove SlotPreference
Previously if we didn't meet the gold count or white counts set in
settings, we would select from other lists that do not match what we
specified in the config. This feels like a side effect and in fact we
should more strictly adher to the prefereces.

If our preferences aren't met, we simply do peer discovery.