hosts: delete darklist entries older than one day
We introduce a new method called refresh() which deletes hosts older
than a specified period. This is applied to the darklist when we load it
on start(), and when we store hosts onto our darklist that are received by
other peers.

We do this to avoid the risk of passing faulty nodes around the network,
and assume that one day is sufficient for valid hosts to propagate to
hosts that support that transport.