net: remove fragile unwrap() from move_host()
Previously, we called unwrap() inside of move_host, which would panic if
the host state is invalid (for example, if we try to move a host that is
currently moving).

This is fragile in async conditions such as the following:

* We are doing a version exchange with a peer called Bob
* Version exchange passes and we move the host to Goldlist
* Bob disconnects while this is ongoing
* We try to move to Greylist -> panic!

Now we simply return an error on the second move, so what should happen
instead is:

* Version exchange passes and we move the host to Goldlist
* Bob disconnects
* We try to move to Greylist -> the move fails
* Bob is still on our Goldlist

This means that Bob's node might be on the goldlist even though it's not
actually online. This will be cleaned up as soon as outbound session
fails to establish a connection to it.